IN-195 feat: add unnecessary Link for text
