--- src/common/components/textUnfolding/TextUnfolding.stories.ts
@@ -1,47 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react'
-
-import { TextUnfolding } from '@/common'
-
-const meta = {
-  args: {
-    charactersToShow: 250,
-  },
-  component: TextUnfolding,
-  parameters: {
-    layout: 'centered',
-  },
-  tags: ['autodocs'],
-  title: 'Component/TextUnfolding',
-} satisfies Meta<typeof TextUnfolding>
-
-export default meta
-type Story = StoryObj<typeof meta>
-
-export const LongTextUnfolding: Story = {
-  args: {
-    children:
-      'Эстетическое воздействие редуцирует диалогический парафраз. Модальность высказывания семантически притягивает речевой акт. Декодирование активно. Мифопорождающее текстовое устройство начинает диссонансный палимпсест, однако дальнейшее развитие приемов декодирования мы находим в работах академика В.Виноградова. Модальность высказывания отталкивает метафоричный метр, таким образом постепенно смыкается с сюжетом. Диалогический контекст отражает дольник.\n' +
-      '\n' +
-      'Холодный цинизм приводит метр. Абстрактное высказывание семантически притягивает словесный амфибрахий. Синхрония представляет собой метафоричный голос персонажа.\n' +
-      '\n' +
-      'Лексика доступна. Ударение, соприкоснувшись в чем-то со своим главным антагонистом в постструктурной поэтике, пространственно неоднородно. Парафраз прекрасно нивелирует мифопоэтический хронотоп, первым образцом которого принято считать книгу А.Бертрана "Гаспар из тьмы". Филологическое суждение интегрирует диалогический контекст. Например, лес — для опытного лесника, охотника, просто внимательного грибника — неисчерпаемое природное семиотическое пространство — текст, поэтому цезура откровенна. Кроме того, постоянно воспроизводится постулат о письме как о технике, обслуживающей язык, поэтому стих вызывает символ.',
-  },
-}
-
-export const ShortTextUnfolding: Story = {
-  args: {
-    children: 'Short text without button',
-  },
-}
-
-export const LongTextInBox: Story = {
-  args: {
-    children:
-      'Эстетическое воздействие редуцирует диалогический парафраз. Модальность высказывания семантически притягивает речевой акт. Декодирование активно. Мифопорождающее текстовое устройство начинает диссонансный палимпсест, однако дальнейшее развитие приемов декодирования мы находим в работах академика В.Виноградова. Модальность высказывания отталкивает метафоричный метр, таким образом постепенно смыкается с сюжетом. Диалогический контекст отражает дольник.\n' +
-      '\n' +
-      'Холодный цинизм приводит метр. Абстрактное высказывание семантически притягивает словесный амфибрахий. Синхрония представляет собой метафоричный голос персонажа.\n' +
-      '\n' +
-      'Лексика доступна. Ударение, соприкоснувшись в чем-то со своим главным антагонистом в постструктурной поэтике, пространственно неоднородно. Парафраз прекрасно нивелирует мифопоэтический хронотоп, первым образцом которого принято считать книгу А.Бертрана "Гаспар из тьмы". Филологическое суждение интегрирует диалогический контекст. Например, лес — для опытного лесника, охотника, просто внимательного грибника — неисчерпаемое природное семиотическое пространство — текст, поэтому цезура откровенна. Кроме того, постоянно воспроизводится постулат о письме как о технике, обслуживающей язык, поэтому стих вызывает символ.',
-    className: 'w-[300px] border rounded-sm p-6',
-  },
-}
